Repository 작성 중
- 파일 업로드 작성하기 전에 임시 저장

+++ release/src/main/java/com/project/release/domain/album/Album.java
@@ -15,10 +15,10 @@ public class Album {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "album_id")
-    private Integer albumId;
+    private Long albumId;
 
     @Column(name = "user_id")
-    private Integer userId;
+    private Long userId;
 
     private String thumbnail;
